+++ src/main/java/com/nba/tpwebservice/service/JoueurService.java
@@ -2,7 +2,6 @@ package com.nba.tpwebservice.service;
 
 import com.nba.tpwebservice.entity.ActionEntity;
 import com.nba.tpwebservice.entity.VActionParJoueurParSaisonEntity;
-import com.nba.tpwebservice.entity.VMatchJoueurEntity;
 import com.nba.tpwebservice.repository.ActionRepository;
 import com.nba.tpwebservice.repository.VActionParJoueurParSaisonRepository;
 import com.nba.tpwebservice.repository.VMatchJoueurRepository;
@@ -25,12 +24,12 @@ public class JoueurService {
     VActionParJoueurParSaisonRepository vActionParJoueurParSaisonRepository;
     @Autowired
     VMatchJoueurRepository vMatchJoueurRepository;
-    public HashMap<String, Integer> getStats(String idSaison, String idJoueur, List<VActionParJoueurParSaisonEntity> actionsSaison, int nombre){
+    public HashMap<String, Float> getStats(String idSaison, String idJoueur, List<VActionParJoueurParSaisonEntity> actionsSaison, int nombre){
         List<ActionEntity> actions = this.actionRepository.findAll();
-        HashMap<String, Integer> hashMapAction = new HashMap<>();
+        HashMap<String, Float> hashMapAction = new HashMap<>();
         System.out.println("saison "+idSaison+" joueur"+idJoueur);
 
-        hashMapAction.put("match", nombre);
+        hashMapAction.put("match", (float) nombre);
         for (ActionEntity action: actions) {
             hashMapAction.put(action.getNom(), (this.getActionState(idSaison, idJoueur, action.getIdAction(), actionsSaison)/nombre));
         }
@@ -40,7 +39,7 @@ public class JoueurService {
     public int getActionState(){
         return 0;
     }
-    public int getActionState(String idSaison, String idJoueur, String idAction, List<VActionParJoueurParSaisonEntity> actions){
+    public float getActionState(String idSaison, String idJoueur, String idAction, List<VActionParJoueurParSaisonEntity> actions){
         int retour = 0;
         for (VActionParJoueurParSaisonEntity action:
              actions) {
